feat: add consistent widget panels with toggle
