+++ app/static/css/style.css
@@ -299,25 +299,34 @@ li:hover {
   transition: width 0.3s ease;
 }
 
-#doc-panel-wrapper.collapsed {
+#doc-panel-wrapper.collapsed,
+#chat-panel-wrapper.collapsed,
+#chat-history-wrapper.collapsed,
+#search-panel-wrapper.collapsed {
   width: 2.5rem;
   flex: 0 0 2.5rem;
   overflow: hidden;
   max-width: 2.5rem;
 }
 
-#doc-panel-wrapper.collapsed .column {
-  padding: 0.5rem;
-}
-
 #doc-panel-wrapper.collapsed h2,
 #doc-panel-wrapper.collapsed form,
 #doc-panel-wrapper.collapsed ul,
-#doc-panel-wrapper.collapsed .status-text {
+#doc-panel-wrapper.collapsed .status-text,
+#chat-panel-wrapper.collapsed h2,
+#chat-panel-wrapper.collapsed form,
+#chat-panel-wrapper.collapsed .chat-box,
+#chat-panel-wrapper.collapsed > div:not(.header-row),
+#chat-history-wrapper.collapsed h2,
+#chat-history-wrapper.collapsed #chat-history-list,
+#search-panel-wrapper.collapsed h2,
+#search-panel-wrapper.collapsed form,
+#search-panel-wrapper.collapsed ul,
+#search-panel-wrapper.collapsed .status-text {
   display: none;
 }
 
-#toggle-docs-btn {
+.toggle-panel-btn {
   font-size: 0.8rem;
   padding: 0.25rem 0.5rem;
   margin-left: auto;
@@ -412,56 +421,24 @@ input:disabled, button:disabled {
 .hidden {
   display: none;
 }
-
-#search-panel-wrapper.collapsed {
-  width: 2.5rem;
-  flex: 0 0 2.5rem;
-  overflow: hidden;
-}
-
-#search-panel-wrapper.collapsed .column {
-  padding: 0.5rem;
-}
-
-#search-panel-wrapper.collapsed h2,
-#search-panel-wrapper.collapsed form,
-#search-panel-wrapper.collapsed ul,
-#search-panel-wrapper.collapsed .status-text {
-  display: none;
-}
-#chat-history {
-  padding: 10px;
-  font-family: var(--font-ui);
-  color: var(--text-main);
-  background-color: var(--bg-panel);
-  border-left: 1px solid var(--border-mid);
-  overflow-y: auto;
-}
-
-#chat-history h2 {
-  margin-bottom: 10px;
-  font-size: 1.1rem;
-  color: var(--text-main);
-}
-
 #chat-history-list {
   list-style: none;
   padding: 0;
   margin: 0;
 }
 
-#chat-history-list li {
-  background: var(--bg-card);
-  margin-bottom: 8px;
-  padding: 10px;
-  border-radius: var(--radius);
+#chat-history-list .session-item {
+  padding: 8px;
+  margin-bottom: 6px;
+  background-color: var(--bg-card);
+  border-radius: 4px;
   border: 1px solid var(--border-mid);
   cursor: pointer;
-  transition: background 0.2s ease;
+  transition: background-color 0.2s ease;
 }
 
-#chat-history-list li:hover {
-  background: var(--bg-hover);
+#chat-history-list .session-item:hover {
+  background-color: var(--bg-hover);
 }
 
 #chat-history-list .session-title {
@@ -469,49 +446,9 @@ input:disabled, button:disabled {
   color: var(--text-main);
 }
 
-#chat-history-list .session-meta {
-  font-size: 0.85em;
-  color: var(--text-dim);
-  margin-top: 4px;
-}
-
-/* --- Chat History Panel --- */
-#chat-history-panel {
-  background-color: #1c1c1c;
-  border-left: 1px solid #444;
-  color: #eee;
-  padding: 10px;
-  width: 280px;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  z-index: 100;
-  overflow-y: auto;
-}
-
-#chat-history-panel .panel-header {
-  font-size: 1.2em;
-  margin-bottom: 10px;
-  border-bottom: 1px solid #444;
-  padding-bottom: 6px;
-}
-
-#chat-history-list .session-item {
-  padding: 8px;
-  margin-bottom: 6px;
-  background-color: #222;
-  border-radius: 4px;
-  cursor: pointer;
-  transition: background-color 0.2s ease;
-}
-
-#chat-history-list .session-item:hover {
-  background-color: #333;
-}
-
 #chat-history-list .session-timestamp {
   font-size: 0.9em;
-  color: #aaa;
+  color: var(--text-dim);
 }
 
 #chat-history-list .loading,
