+++ app/static/js/chatHistory.js
@@ -1,36 +1,9 @@
 // Codex: Do NOT load backend or Python files. This file is frontend-only.
 // chatHistory.js
 
-import { $ } from './dom.js';
+import { $, initPanelToggle } from './dom.js';
 import { setSessionId } from './session.js';
-
-/**
- * Clears the chat UI.
- */
-function clearChatBox() {
-  const chatBox = $("chat-box");
-  if (chatBox) chatBox.innerHTML = '';
-  const searchResults = $("search-results");
-  if (searchResults) searchResults.innerHTML = '';
-}
-
-/**
- * Renders a single message pair to the chat box.
- * @param {string} userText 
- * @param {string} assistantText 
- */
-function renderMessagePair(userText, assistantText) {
-  const chatBox = $("chat-box");
-  if (!chatBox) return;
-
-  const userDiv = document.createElement("div");
-  userDiv.innerHTML = `<strong>You:</strong> ${userText}`;
-  chatBox.appendChild(userDiv);
-
-  const botDiv = document.createElement("div");
-  botDiv.innerHTML = `<strong>Assistant:</strong> ${assistantText}`;
-  chatBox.appendChild(botDiv);
-}
+import { renderMessagePair, clearChatUI, renderChatHistoryList } from './render.js';
 
 /**
  * Loads and displays messages for a specific session.
@@ -44,21 +17,16 @@ async function loadSessionMessages(sessionId) {
     const data = await res.json();
     const history = data.history;
 
-    clearChatBox();
+    clearChatUI();
 
     for (const [user, assistant] of history) {
       renderMessagePair(user, assistant);
     }
   } catch (err) {
     console.error("❌ Error loading session messages:", err);
-    clearChatBox();
+    clearChatUI();
     renderMessagePair("System", `⚠️ Failed to load session history: ${err.message}`);
   }
-}
-
-function formatTimestamp(isoString) {
-  const date = new Date(isoString);
-  return date.toLocaleString();
 }
 
 /**
@@ -70,6 +38,8 @@ export const chatHistory = {
     const container = $("chat-history-list");
     if (!container) return;
 
+    initPanelToggle('chat-history-wrapper', 'toggle-history-btn');
+
     container.innerHTML = `<div class="loading">Loading sessions...</div>`;
 
     fetch("/sessions")
@@ -78,22 +48,11 @@ export const chatHistory = {
         return res.json();
       })
       .then(sessions => {
-        container.innerHTML = "";
-
-        sessions.forEach(sessionEntry => {
-          const div = document.createElement("div");
-          div.className = "session-item";
-          div.innerHTML = `
-            <div class="session-title"><strong>Session ID:</strong> ${sessionEntry.session_id}</div>
-            <div class="session-timestamp">${formatTimestamp(sessionEntry.created_at)}</div>
-          `;
-          div.addEventListener("click", () => {
-            console.log("📦 Loading session:", sessionEntry.session_id);
-            setSessionId(sessionEntry.session_id);
-            session.sessionId = sessionEntry.session_id;
-            loadSessionMessages(sessionEntry.session_id);
-          });
-          container.appendChild(div);
+        renderChatHistoryList(container, sessions, (sessionId) => {
+          console.log("📦 Loading session:", sessionId);
+          setSessionId(sessionId);
+          session.sessionId = sessionId;
+          loadSessionMessages(sessionId);
         });
       })
       .catch(err => {
