+++ app/templates/index.html
@@ -45,19 +45,20 @@
 
     <div class="feature-layout">
       <div class="container">
-        <div id="doc-panel-wrapper" class="column-wrapper scrollable">
+        <div class="column scrollable">
 
-          <!-- ✅ CHAT HISTORY PANEL -->
-          <div id="chat-history-panel">
-            <div class="panel-header">Chat History</div>
+          <div id="chat-history-wrapper" class="column-wrapper">
+            <div class="header-row">
+              <h2>Chat History</h2>
+              <button id="toggle-history-btn" class="toggle-panel-btn" title="Minimize Chat History" type="button">«</button>
+            </div>
             <div id="chat-history-list" class="loading">Loading...</div>
           </div>
 
-          <!-- ✅ DOCUMENT LIBRARY -->
-          <div class="column" id="doc-manager">
+          <div id="doc-panel-wrapper" class="column-wrapper">
             <div class="header-row">
               <h2>Document Library</h2>
-              <button id="toggle-docs-btn" title="Minimize Document Library" type="button">«</button>
+              <button id="toggle-docs-btn" class="toggle-panel-btn" title="Minimize Document Library" type="button">«</button>
             </div>
 
             <form id="upload-form" method="post" enctype="multipart/form-data">
@@ -78,33 +79,40 @@
           </div>
         </div>
 
-        <!-- CHAT + SEARCH -->
         <div class="column scrollable">
-          <h2>Assistant Chat</h2>
-          <div id="chat-box" class="chat-box">
-            <div>
-              <strong>Assistant:</strong> Hello! How can I help you today?
+
+          <div id="chat-panel-wrapper" class="column-wrapper">
+            <div class="header-row">
+              <h2>Assistant Chat</h2>
+              <button id="toggle-chat-btn" class="toggle-panel-btn" title="Minimize Chat" type="button">«</button>
+            </div>
+            <div id="chat-box" class="chat-box">
+              <div>
+                <strong>Assistant:</strong> Hello! How can I help you today?
+              </div>
+            </div>
+
+            <form id="chat-form">
+              <input type="text" id="user-input" placeholder="Type a message..." />
+              <button type="submit" id="submit-button">Send</button>
+            </form>
+
+            <div style="display: flex; align-items: center; gap: 1em; margin-bottom: 1em;">
+              <button id="open-persona-btn">Edit Persona</button>
+              <button id="new-chat" type="button">New Chat</button>
+              <button id="clear-history" type="button">Clear Logs</button>
+              <button id="reset-llm" type="button">Reset LLM</button>
+              <button id="download-chat">Download Chat</button>
+              <input type="file" id="upload-chat-file" accept=".html" style="display: none" />
+              <button id="upload-chat-btn" type="button">Upload Chat</button>
             </div>
           </div>
 
-          <form id="chat-form">
-            <input type="text" id="user-input" placeholder="Type a message..." />
-            <button type="submit" id="submit-button">Send</button>
-          </form>
-
-          <div style="display: flex; align-items: center; gap: 1em; margin-bottom: 1em;">
-            <button id="open-persona-btn">Edit Persona</button>
-            <button id="new-chat" type="button">New Chat</button>
-            <button id="clear-history" type="button">Clear Logs</button>
-            <button id="reset-llm" type="button">Reset LLM</button>
-            <button id="download-chat">Download Chat</button>
-            <input type="file" id="upload-chat-file" accept=".html" style="display: none" />
-            <button id="upload-chat-btn" type="button">Upload Chat</button>
-          </div>
-
           <div id="search-panel-wrapper" class="column-wrapper">
-            <button id="toggle-search-btn" title="Minimize Search Results">🡸</button>
-            <h2>Search Documents</h2>
+            <div class="header-row">
+              <h2>Search Documents</h2>
+              <button id="toggle-search-btn" class="toggle-panel-btn" title="Minimize Search Results" type="button">«</button>
+            </div>
             <form id="search-form">
               <input type="text" id="search-query" placeholder="Enter search text..." />
               <label for="top-k-select">Number of results:</label>
@@ -117,6 +125,7 @@
             </form>
             <div id="search-results"></div>
           </div>
+
         </div>
       </div>
     </div>
